Fix namespace issues and type references - WIP
- Add Common namespace for PACKET_FIX/PACKET_VAR to 100+ packet files
- Fix DamageResult ambiguity (rename to AttackResult in ICombatService)
- Fix IPlayerClient namespace in 33+ handler files (contracts.Interfaces)
- Fix SLOT_MOVE to MOVE_SLOT in InventoryService
- Fix XmlSubjectList to SkillSubjectList in HealingService
- Add GetEntityTypeId() to Monster.cs for IEntity interface
- Add missing PacketType entries (CS_DISTRIBUTE_COMBAT_JOB_STATUS/SKILL_POINT)
- Delete duplicate Combat/DamageResult.cs

Build still has errors - checkpoint commit for progress

+++ kakia_lime_odyssey_server/Combat/DamageHandler.cs
@@ -114,7 +114,7 @@ public static class DamageHandler
 			subIsMiss = rnd.Next(0, 100) >= subHitChance;
 
 			// Off-hand critical
-			int subCritChance = Math.Clamp(sourceStatus.SubAttack.CritRate, 0, 100);
+			int subCritChance = Math.Clamp((int)sourceStatus.SubAttack.CritRate, 0, 100);
 			subIsCrit = !subIsMiss && rnd.Next(0, 100) < subCritChance;
 
 			// Off-hand damage (50% penalty)
